benerin customer
- customer security tested jalan
- tambah kafka listener di customer buat nambah asset coin
- nambah client keycloak "super-client" yg punya privilege ke all scopes (buat testing biar gampang)
- ubah port order service ke 8082 biar ga tabrakan sama keycloak

+++ customer-service/src/main/java/com/achsat/crypto/customer/model/CustomerAssets.java
@@ -17,10 +17,10 @@ public class CustomerAssets {
     private Integer id;
 
     @Column(name="customer_id")
-    private Integer customerId;
+    private Long customerId;
 
     @Column(name="coin_id")
-    private Integer coinId;
+    private Long coinId;
 
     @Column(name="qty")
     private BigDecimal qty;
